feat: show toast for rpc errors
